Relase MVP version

+++ server/modules/wallets/model/withdraw.go
@@ -2,5 +2,5 @@ package walletmodel
 
 type Withdraw struct {
 	Tx     string `json:"tx"`
-	Amount int    `json:"amount"`
+	Amount int64  `json:"amount"`
 }
